fixes #590098
Attempting to build media-libs/libraw on the musl libc tickles a bug in vect_analyze_data_ref_accesses on GCC 4.9 and 5.4 (only ones I tested, apparently 5.3 is also affected).

The attached patch, from upstream, resolves this issue.  As upstream seems to be unwilling to push out a new GCC 4.9 version, and 5.4 did not include this particular fix, it may be beneficial to carry this patch in the Gentoo patch set.