dev-lang/rust-bin: add initial ebuild
this is a little ebuild I've been using for a while, it's a copycat
of the rust-bin ebuild from the tree, obviously. It's far from being
perfect, but it's a massive benefit for low end machines.

On top of that, more up to date =dev-lang/rust-1.38.0 is a formidable
bug festival, cargo build is broken by libgit2 issues.

Signed-off-by: Steffen Kuhn <nielson2@yandex.com>